Tests/LibWeb/Text/expected/HTML/form-element-enctype.txt

   enctype  | application/x-www-form-urlencoded

enctype  | text/plain

enctype  | application/x-www-form-urlencoded
